LF-17090-10 drm/bridge: synopsys: dw-mipi-dsi2: Get PHY escape clk rate from phy_ops
PHY escape clk rate should be SoC vendor dependent since the PHY vendor
and/or the escape clk's parent could be different.  So, get PHY escape
clk rate from phy_ops->get_esc_clk_rate() callback.  While at it, change
phy_ops->get_esc_clk_rate() callback's parameter esc_clk_rate type from
"unsigned int *" to "unsigned long *" as a clock rate in Hz is more
accurate and is usually in type "unsigned long".

+++ include/drm/bridge/dw_mipi_dsi2.h
@@ -58,7 +58,7 @@ struct dw_mipi_dsi2_phy_ops {
 			     unsigned int *lane_mbps);
 	int (*get_timing)(void *priv_data, unsigned int lane_mbps,
 			  struct dw_mipi_dsi2_phy_timing *timing);
-	int (*get_esc_clk_rate)(void *priv_data, unsigned int *esc_clk_rate);
+	int (*get_esc_clk_rate)(void *priv_data, unsigned long *esc_clk_rate);
 };
 
 struct dw_mipi_dsi2_host_ops {
